feat: 通信プロトコルとコマンド処理の実装
- ドライブ、安全状態、自動コマンドストア、送信ポートへの参照を保持するコンテキスト構造体を追加。
- 登録とディスパッチ機能を備えたコマンドハンドラ管理用のディスパッチャを導入。
- ACKおよびステータスメッセージの送信処理を行う送信ポートクラスを作成。
- 自動モード、自動設定値、ハートビート、キル、クリアキルコマンド用のハンドラを開発。
- COBS符号化とCRCチェックを用いたパケットの読み書きを実装。
- メッセージタイプとペイロード構造を含むプロトコル定義を確立。
- シリアル通信設定とハードウェアパラメータの設定を統合。
- 設定定数を利用するためDriveクラスとEngineクラスをリファクタリング。
- UART通信と制御ロジックを処理するメインアプリケーションループを追加。

+++ firmware/src/hardware/Drive.h
@@ -13,7 +13,7 @@ public:
 	void control();
 
 	void setSpeed(int newSpeed);
-	void setAngle(int newAngle);
+	void setAngle(float newAngle);
 
 	void updateTimeout();
 	bool evalTimeout();
